First iteration on increasing UT coverage - simple tests / missing files mostly
