+++ client/src/components/Card/Settings/index.tsx
@@ -1,6 +1,6 @@
 import Settings from './Settings';
-import SettingsActions from './SettingsActions';
-import SettingsStore from './SettingsStore';
+import SettingsActions from './CardSettingsActions';
+import SettingsStore from './CardSettingsStore';
 
 export {
   Settings,
