+++ client/src/components/Navbar/index.tsx
@@ -6,7 +6,6 @@ import FontIcon from 'react-md/lib/FontIcons';
 import ListItem from 'react-md/lib/Lists/ListItem';
 import Avatar from 'react-md/lib/Avatars';
 import SelectField from 'react-md/lib/SelectFields';
-import NavigationLink from './NavigationLink';
 import { Link } from 'react-router';
 import Chip from 'react-md/lib/Chips';
 import Menu from 'react-md/lib/Menus/Menu';
