+++ client/src/components/Setup/index.tsx
@@ -105,11 +105,15 @@ export default class Setup extends React.Component<any, ISetupState> {
   fixRedirectUrl(redirectUrl: string): string {
     if (redirectUrl) { return redirectUrl; }
 
-    let host = window.location.host;
+    let host = '';
+    if (window && window.location) {
+      host = window.location.host;
+    }
+
     let protocol = 'https:';
 
     // On localhost, authentication requests go directly to port 4000
-    if (host === 'localhost:3000' || host === 'localhost:4000') { 
+    if (host === '' || host === 'localhost:3000' || host === 'localhost:4000') { 
       host = 'localhost:4000'; 
       protocol = 'http:';
     }
@@ -176,7 +180,9 @@ export default class Setup extends React.Component<any, ISetupState> {
   }
 
   redirectOut() {
-    window.location.replace('/'); 
+    if (window && window.location) {
+      window.location.replace('/'); 
+    }
   }
 
   onRemoveAdmin(admin: string) {
