+++ client/src/tests/mocks/requests/configuration.ts
@@ -27,6 +27,14 @@ function mockRequests() {
         window.dashboard = dashboard || null;
       })(window);
     `);
+
+    nock('http://localhost')
+    .get('/api/dashboard/id')
+    .reply(200, `
+     
+          return ${JSON.stringify(dashboard)};
+      
+    `);
 }
 export {
   mockRequests
